src/requirements.txt
# Azure Functions Core

azure-functions

# Azure Services
#azure-storage-blob>=12.17.0
#azure-cognitiveservices-speech>=1.34.0
#azure-identity>=1.15.0
#azure-keyvault-secrets>=4.7.0

# HTTP and API
requests>=2.31.0
#aiohttp>=3.9.0

# Audio Processing
#pydub>=0.25.1
#librosa>=0.10.1
#soundfile>=0.12.1

# Text Processing
#nltk>=3.8.1
#beautifulsoup4>=4.12.2
#markdown>=3.5.1

# Data Handling
##pandas>=2.1.4
numpy>=1.24.3

# Logging and Monitoring
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1

# Configuration
#python-dotenv>=1.0.0
#pydantic>=2.5.0

# Utilities
#chardet>=5.2.0
#python-dateutil>=2.8.2
